auth_service/requirements.txt
flask
flask_sqlalchemy
pyjwt
werkzeug
flask-cors